add confirm to shell purge

+++ cmd/plural/shell.go
@@ -65,7 +65,13 @@ func handleShellSync(c *cli.Context) error {
 	return handleUnlock(c)
 }
 
+var destoryShellConfirm = "Are you sure you want to destroy your cloud shell (you should either `plural destroy` anything deployed or `plural shell sync` to sync the contents locally)?"
+
 func handleShellPurge(c *cli.Context) error {
+	if ok := confirm(destoryShellConfirm); !ok {
+		return nil
+	}
+
 	client := api.NewClient()
 	return client.DeleteShell()
 }
